+++ src/Query/AverageField.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Aternos\Model\Query;
+
+/**
+ * Class AverageField
+ *
+ * @package Aternos\Model\Query
+ */
+class AverageField extends SelectField
+{
+    public ?int $function = self::AVERAGE;
+}