+++ src/Query/SumField.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Aternos\Model\Query;
+
+/**
+ * Class SumField
+ *
+ * @package Aternos\Model\Query
+ */
+class SumField extends SelectField
+{
+    public ?int $function = self::SUM;
+}