+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   query functions will automatically filter the results to the correct variant. Calls to the shared
   parent model will return all variants but of the correct subtype.
 - Added `CountField` and `GenericModel::count()` for easier count queries.
+- Added `SumField` and `AverageField`.
 
 ### Removed
 - Removed `SimpleModel`.